Move board-related classes to src/boards and tests/boards. (#2147)
