+++ src/Game.ts
@@ -2,8 +2,8 @@ import * as constants from './constants';
 import {ALL_CORPORATION_DECKS} from './cards/AllCards';
 import {AndOptions} from './inputs/AndOptions';
 import {BeginnerCorporation} from './cards/corporation/BeginnerCorporation';
-import {BoardColony, Board} from './Board';
-import {BoardName} from './BoardName';
+import {BoardColony, Board} from './boards/Board';
+import {BoardName} from './boards/BoardName';
 import {CardFinder} from './CardFinder';
 import {CardName} from './CardName';
 import {CardType} from './cards/CardType';
@@ -17,9 +17,9 @@ import {CorporationCard} from './cards/corporation/CorporationCard';
 import {Database} from './database/Database';
 import {Dealer} from './Dealer';
 import {Decks} from './Deck';
-import {ElysiumBoard} from './ElysiumBoard';
+import {ElysiumBoard} from './boards/ElysiumBoard';
 import {FundedAward} from './FundedAward';
-import {HellasBoard} from './HellasBoard';
+import {HellasBoard} from './boards/HellasBoard';
 import {IAward} from './awards/IAward';
 import {ISerializable} from './ISerializable';
 import {IMilestone} from './milestones/IMilestone';
@@ -31,7 +31,7 @@ import {LogHelper} from './components/LogHelper';
 import {LogMessage} from './LogMessage';
 import {ORIGINAL_MILESTONES, VENUS_MILESTONES, ELYSIUM_MILESTONES, HELLAS_MILESTONES, ARES_MILESTONES} from './milestones/Milestones';
 import {ORIGINAL_AWARDS, VENUS_AWARDS, ELYSIUM_AWARDS, HELLAS_AWARDS, ARES_AWARDS} from './awards/Awards';
-import {OriginalBoard} from './OriginalBoard';
+import {OriginalBoard} from './boards/OriginalBoard';
 import {PartyHooks} from './turmoil/parties/PartyHooks';
 import {Phase} from './Phase';
 import {Player, PlayerId} from './Player';
