+++ src/awards/EstateDealer.ts
@@ -1,7 +1,7 @@
 import {IAward} from './IAward';
 import {Player} from '../Player';
 import {Game} from '../Game';
-import {Board} from '../Board';
+import {Board} from '../boards/Board';
 
 export class EstateDealer implements IAward {
     public name: string = 'Estate Dealer';
