+++ src/cards/corporation/TharsisRepublic.ts
@@ -7,7 +7,7 @@ import {SpaceType} from '../../SpaceType';
 import {ISpace} from '../../ISpace';
 import {Resources} from '../../Resources';
 import {CardName} from '../../CardName';
-import {Board} from '../../Board';
+import {Board} from '../../boards/Board';
 import {CardType} from '../CardType';
