+++ tests/TestingUtils.ts
@@ -2,7 +2,7 @@ import {Player} from '../src/Player';
 import {Game, GameOptions} from '../src/Game';
 import * as constants from '../src/constants';
 import {SpaceType} from '../src/SpaceType';
-import {BoardName} from '../src/BoardName';
+import {BoardName} from '../src/boards/BoardName';
 import {RandomMAOptionType} from '../src/RandomMAOptionType';
 import {ISpace} from '../src/ISpace';
 import {Color} from '../src/Color';
