+++ tests/boards/Board.spec.ts
@@ -1,10 +1,10 @@
 import {expect} from 'chai';
-import {OriginalBoard} from '../src/OriginalBoard';
-import {Player} from '../src/Player';
-import {TileType} from '../src/TileType';
-import {ISpace} from '../src/ISpace';
-import {SpaceType} from '../src/SpaceType';
-import {TestPlayers} from './TestingUtils';
+import {OriginalBoard} from '../../src/boards/OriginalBoard';
+import {Player} from '../../src/Player';
+import {TileType} from '../../src/TileType';
+import {ISpace} from '../../src/ISpace';
+import {SpaceType} from '../../src/SpaceType';
+import {TestPlayers} from '../TestingUtils';
 
 describe('Board', function() {
   let board : OriginalBoard; let player : Player; let player2 : Player;
